Add support for 32-bit Cues.
MaestroController - Fixed Section Cue check.
PaletteController - Fixed `find()`.
CueInterpreterDialog - Added close button, minor UI tweaks.
CueModel - Added numbering to C++ byte arrays.
AnimationControlWidget - Fixed auto-generated name added to unrecognized
Palettes.

+++ src/model/cuemodel.h
@@ -12,10 +12,10 @@
 namespace PixelMaestroStudio {
 	class CueModel : public QStandardItemModel {
 		public:
-			CueModel(uint8_t* cue, uint16_t size);
+			CueModel(uint8_t* cue, uint32_t size);
 
 		private:
-			int add_cue(uint8_t* cue, uint16_t size);
+			int add_cue(uint8_t* cue, uint32_t size);
 	};
 }
